arreglo un bug q impedia regenerar el informe html

+++ plasTeX/Renderers/PageTemplate/__init__.py
@@ -191,30 +191,35 @@ def copytree(src, dest, symlink=None):
 
     """
     for root, dirs, files in os.walk(src, True):
+      if not ('.svn' in root):
         for d in dirs:
-            srcpath = os.path.join(root, d)
-            destpath = os.path.join(dest, root, d)
-            if symlink and os.path.islink(srcpath):
-                if os.path.exists(destpath):
-                    os.remove(destpath)
-                os.symlink(os.readlink(srcpath), destpath)
-            elif not os.path.isdir(destpath):
-                os.makedirs(destpath)
-                try: 
-                    shutil.copymode(srcpath, destpath)
-                except: pass
-                try: 
-                    shutil.copystat(srcpath, destpath)
-                except: pass
+            if not ('.svn' in d):
+                
+                srcpath = os.path.join(root, d)
+                destpath = os.path.join(dest, root, d)
+                if symlink and os.path.islink(srcpath):
+                    if os.path.exists(destpath):
+                        os.remove(destpath)
+                    os.symlink(os.readlink(srcpath), destpath)
+                elif not os.path.isdir(destpath):
+                    os.makedirs(destpath)
+                    try: 
+                        shutil.copymode(srcpath, destpath)
+                    except: pass
+                    try: 
+                        shutil.copystat(srcpath, destpath)
+                    except: pass
         for f in files:
-            srcpath = os.path.join(root, f)
-            destpath = os.path.join(dest, root, f)
-            if symlink and os.path.islink(srcpath):
-                if os.path.exists(destpath):
-                    os.remove(destpath)
-                os.symlink(os.readlink(srcpath), destpath)
-            else:
-                shutil.copy2(srcpath, destpath)
+            if not ('.svn' in f):
+                
+                srcpath = os.path.join(root, f)
+                destpath = os.path.join(dest, root, f)
+                if symlink and os.path.islink(srcpath):
+                    if os.path.exists(destpath):
+                        os.remove(destpath)
+                    os.symlink(os.readlink(srcpath), destpath)
+                else:
+                    shutil.copy2(srcpath, destpath)
 
 class TemplateEngine(object):
     def __init__(self, ext, function):
